@@ -5,26 +5,28 @@
   "main": "src/index.js",
   "proxy": "http://backend",
   "dependencies": {
-    "antd": "^3.6.2",
-    "chroma-js": "^1.3.7",
-    "date-fns": "^1.29.0",
-    "libphonenumber-js": "^1.2.18",
+    "ajv": "^8.11.0",
+    "antd": "^3.26.20",
+    "chroma-js": "^1.4.1",
+    "date-fns": "^1.30.1",
+    "libphonenumber-js": "^1.10.6",
     "lodash.isequal": "^4.5.0",
-    "mobx": "^5.0.3",
-    "mobx-react": "^5.2.3",
-    "mobx-state-tree": "^3.0.2",
-    "react": "^16.4.1",
-    "react-dom": "^16.4.1",
-    "react-lazy-load": "^3.0.13",
-    "react-loadable": "^5.4.0",
-    "react-pose": "^2.1.2",
+    "mobx": "^5.15.7",
+    "mobx-react": "^5.4.4",
+    "mobx-state-tree": "^3.17.3",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
+    "react-lazy-load": "^3.1.13",
+    "react-loadable": "^5.5.0",
+    "react-pose": "^2.3.0",
     "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1",
-    "react-scripts": "1.1.4",
-    "styled-components": "^3.3.2",
+    "react-scripts": "^4.0.3",
+    "styled-components": "^5.3.5",
     "to-camel-case": "^1.0.0",
     "to-snake-case": "^1.0.0",
-    "traverse": "^0.6.6"
+    "traverse": "^0.6.6",
+    "typescript": "^4.3.5"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -48,7 +50,7 @@
     ]
   },
   "devDependencies": {
-    "react-cosmos": "^4.6.4",
-    "react-cosmos-context-proxy": "^4.6.1"
+    "react-cosmos": "^5.7.2",
+    "react-cosmos-context-proxy": "^4.8.2"
   }
 }
